requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
requests>=2.31.0  # For HTTP requests (EODHD provider)
python-dotenv>=1.0.0  # For loading .env files

# Data providers (choose based on your needs)
yfinance>=0.2.28  # Yahoo Finance data (free)
alpaca-py>=0.9.0  # Alpaca data and trading

# Optional: For plotting and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.22.0
